@ResetTimer$qv must be defined as far.
In XOSL/MISC/Starter.asm, @ResetTimer$qv is declared as far. However, it
wasn't in XOSL/Timer/Rtc.asm. This mismatch meant that when Starter.asm
calls @ResetTimer$qv, it was using a far call which pushes CS:IP onto
the stack. However, since Rtc.asm doesn't declare @ResetTimer$qv as far,
it only performs a near return, thus popping IP off of the stack and not
CS. This eventually ends up preventing the bootloader from returning the
boot drive in the DL register for the chain load.